pyproject.toml
[project]
name = "ar-checks-demo"
version = "0.1.0"
description = "Amazon Bedrock Automated Reasoning Checks Demo and Utilities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.40",
    "openai>=1.109.1",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "strands-agents>=1.10.0",
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
    "vulture>=2.14",
]
